cleanup and improve mapping model coverage analytics
