+++ legend-engine-xt-analytics-mapping-api/src/main/java/org/finos/legend/engine/api/analytics/MappingAnalytics.java
@@ -14,31 +14,33 @@
 
 package org.finos.legend.engine.api.analytics;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
 import io.opentracing.Scope;
 import io.opentracing.util.GlobalTracer;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
-import org.eclipse.collections.api.RichIterable;
 import org.eclipse.collections.api.list.MutableList;
 import org.eclipse.collections.impl.utility.ListIterate;
 import org.finos.legend.engine.api.analytics.model.MappingModelCoverageAnalysisInput;
-import org.finos.legend.engine.api.analytics.model.MappingModelCoverageAnalysisResult;
 import org.finos.legend.engine.api.analytics.model.MappingRuntimeCompatibilityAnalysisInput;
 import org.finos.legend.engine.api.analytics.model.MappingRuntimeCompatibilityAnalysisResult;
 import org.finos.legend.engine.language.pure.compiler.toPureGraph.HelperModelBuilder;
 import org.finos.legend.engine.language.pure.compiler.toPureGraph.HelperRuntimeBuilder;
 import org.finos.legend.engine.language.pure.compiler.toPureGraph.PureModel;
 import org.finos.legend.engine.language.pure.modelManager.ModelManager;
+import org.finos.legend.engine.protocol.analytics.model.MappingModelCoverageAnalysisResult;
 import org.finos.legend.engine.protocol.pure.v1.model.context.PureModelContextData;
 import org.finos.legend.engine.protocol.pure.v1.model.packageableElement.runtime.PackageableRuntime;
+import org.finos.legend.engine.shared.core.ObjectMapperFactory;
 import org.finos.legend.engine.shared.core.api.result.ManageConstantResult;
 import org.finos.legend.engine.shared.core.kerberos.ProfileManagerHelper;
 import org.finos.legend.engine.shared.core.operational.errorManagement.ExceptionTool;
 import org.finos.legend.engine.shared.core.operational.http.InflateInterceptor;
 import org.finos.legend.engine.shared.core.operational.logs.LoggingEventType;
-import org.finos.legend.pure.generated.Root_meta_analytics_mapping_modelCoverage_MappedEntity;
-import org.finos.legend.pure.generated.core_analytics_mapping_modelCoverageAnalytics;
+import org.finos.legend.pure.generated.Root_meta_analytics_mapping_modelCoverage_MappingModelCoverageAnalysisResult;
+import org.finos.legend.pure.generated.core_analytics_mapping_modelCoverage_analytics;
+import org.finos.legend.pure.generated.core_analytics_mapping_modelCoverage_serializer;
 import org.finos.legend.pure.m3.coreinstance.meta.pure.mapping.Mapping;
 import org.pac4j.core.profile.CommonProfile;
 import org.pac4j.core.profile.ProfileManager;
@@ -53,12 +55,11 @@ import javax.ws.rs.QueryParam;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 
-import static org.finos.legend.engine.api.analytics.MappingAnalyticsHelper.buildMappedEntity;
-
 @Api(tags = "Analytics - Model")
 @Path("pure/v1/analytics/mapping")
 public class MappingAnalytics
 {
+    private final ObjectMapper objectMapper = ObjectMapperFactory.getNewStandardObjectMapperWithPureProtocolExtensionSupports();
     private final ModelManager modelManager;
 
     public MappingAnalytics(ModelManager modelManager)
@@ -71,10 +72,10 @@ public class MappingAnalytics
     @ApiOperation(value = "Analyze the mapping to generate information about mapped classes and mapped properties of each class")
     @Consumes({MediaType.APPLICATION_JSON, InflateInterceptor.APPLICATION_ZLIB})
     @Produces(MediaType.APPLICATION_JSON)
-    public Response doModelCoverageAnalytics(MappingModelCoverageAnalysisInput input,
-                                             @QueryParam("returnMappedEntityInfo") @DefaultValue("false") boolean returnMappedEntityInfo,
-                                             @QueryParam("returnMappedPropertyInfo") @DefaultValue("false") boolean returnMappedPropertyInfo,
-                                             @ApiParam(hidden = true) @Pac4JProfileManager ProfileManager<CommonProfile> pm)
+    public Response analyzeModelCoverage(MappingModelCoverageAnalysisInput input,
+                                         @QueryParam("returnMappedEntityInfo") @DefaultValue("false") boolean returnMappedEntityInfo,
+                                         @QueryParam("returnMappedPropertyInfo") @DefaultValue("false") boolean returnMappedPropertyInfo,
+                                         @ApiParam(hidden = true) @Pac4JProfileManager ProfileManager<CommonProfile> pm)
     {
         MutableList<CommonProfile> profiles = ProfileManagerHelper.extractProfiles(pm);
         PureModel pureModel = this.modelManager.loadModel(input.model, input.clientVersion, profiles, null);
@@ -83,8 +84,9 @@ public class MappingAnalytics
         {
             try
             {
-                RichIterable<? extends Root_meta_analytics_mapping_modelCoverage_MappedEntity> mappedEntities = core_analytics_mapping_modelCoverageAnalytics.Root_meta_analytics_mapping_modelCoverage_doModelCoverageAnalytics_Mapping_1__MappedEntity_MANY_(mapping, pureModel.getExecutionSupport());
-                return ManageConstantResult.manageResult(profiles, new MappingModelCoverageAnalysisResult(mappedEntities.collect(e -> buildMappedEntity(e, returnMappedEntityInfo, returnMappedPropertyInfo, pureModel.getExecutionSupport())).toList()));
+                Root_meta_analytics_mapping_modelCoverage_MappingModelCoverageAnalysisResult analysisResult = core_analytics_mapping_modelCoverage_analytics.Root_meta_analytics_mapping_modelCoverage_analyze_Mapping_1__Boolean_1__Boolean_1__MappingModelCoverageAnalysisResult_1_(mapping, returnMappedEntityInfo, returnMappedPropertyInfo, pureModel.getExecutionSupport());
+                MappingModelCoverageAnalysisResult result = this.objectMapper.readValue(core_analytics_mapping_modelCoverage_serializer.Root_meta_analytics_mapping_modelCoverage_serialization_json_getSerializedMappingModelCoverageAnalysisResult_MappingModelCoverageAnalysisResult_1__String_1_(analysisResult, pureModel.getExecutionSupport()), MappingModelCoverageAnalysisResult.class);
+                return ManageConstantResult.manageResult(profiles, result);
             }
             catch (Exception e)
             {
