--- legend-engine-xt-analytics-mapping-api/src/main/java/org/finos/legend/engine/api/analytics/model/MappedPropertyInfo.java
@@ -1,41 +1,0 @@
-// Copyright 2022 Goldman Sachs
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package org.finos.legend.engine.api.analytics.model;
-
-import org.finos.legend.engine.protocol.pure.v1.model.packageableElement.domain.Multiplicity;
-
-public class MappedPropertyInfo
-{
-    public Multiplicity multiplicity;
-    public MappedPropertyType propertyType;
-
-    public MappedPropertyInfo(Multiplicity multiplicity, MappedPropertyType propertyType)
-    {
-        this.multiplicity = multiplicity;
-        this.propertyType = propertyType;
-    }
-
-    public MappedPropertyInfo(Multiplicity multiplicity)
-    {
-        this.multiplicity = multiplicity;
-        this.propertyType = null;
-    }
-
-    public MappedPropertyInfo(MappedPropertyType propertyType)
-    {
-        this.multiplicity = null;
-        this.propertyType = propertyType;
-    }
-}
