+++ legend-engine-xt-analytics-mapping-api/src/test/java/org/finos/legend/engine/api/analytics/test/TestMappingAnalyticsApi.java
@@ -16,6 +16,7 @@ package org.finos.legend.engine.api.analytics.test;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.finos.legend.engine.api.analytics.MappingAnalytics;
+import org.finos.legend.engine.api.analytics.model.MappingModelCoverageAnalysisInput;
 import org.finos.legend.engine.api.analytics.model.MappingRuntimeCompatibilityAnalysisInput;
 import org.finos.legend.engine.language.pure.modelManager.ModelManager;
 import org.finos.legend.engine.protocol.pure.v1.model.context.PureModelContextData;
@@ -31,6 +32,24 @@ public class TestMappingAnalyticsApi
 {
     private final MappingAnalytics api = new MappingAnalytics(new ModelManager(DeploymentMode.TEST));
     private static final ObjectMapper objectMapper = ObjectMapperFactory.getNewStandardObjectMapperWithPureProtocolExtensionSupports();
+
+    @Test
+    public void testMappingModelCoverageAnalysis() throws IOException
+    {
+        PureModelContextData modelContextData = objectMapper.readValue(Objects.requireNonNull(getClass().getClassLoader().getResource("mappingModelCoverageAnalysisTestData.json")), PureModelContextData.class);
+        Assert.assertEquals(
+                "{\"mappedEntities\":[{\"path\":\"model::Bank\",\"properties\":[{\"_type\":\"entity\",\"entityPath\":\"model::Trader\",\"name\":\"employees\"},{\"_type\":\"enum\",\"enumPath\":\"model::OrgType\",\"name\":\"type\"}]},{\"path\":\"model::Trader\",\"properties\":[{\"_type\":\"MappedProperty\",\"name\":\"fullName\"}]}]}",
+                api.analyzeModelCoverage(new MappingModelCoverageAnalysisInput("vX_X_X", "model::mapping", modelContextData), false, false, null).getEntity().toString());
+    }
+
+    @Test
+    public void testMappingModelCoverageAnalysisWithTypeInfoReturned() throws IOException
+    {
+        PureModelContextData modelContextData = objectMapper.readValue(Objects.requireNonNull(getClass().getClassLoader().getResource("mappingModelCoverageAnalysisTestData.json")), PureModelContextData.class);
+        Assert.assertEquals(
+                "{\"mappedEntities\":[{\"info\":{\"isRootEntity\":true,\"subClasses\":[]},\"path\":\"model::Bank\",\"properties\":[{\"_type\":\"entity\",\"entityPath\":\"model::Trader\",\"mappedPropertyInfo\":{\"multiplicity\":{\"lowerBound\":0},\"propertyType\":\"Entity\"},\"name\":\"employees\"},{\"_type\":\"enum\",\"enumPath\":\"model::OrgType\",\"mappedPropertyInfo\":{\"multiplicity\":{\"lowerBound\":1,\"upperBound\":1},\"propertyType\":\"Enumeration\"},\"name\":\"type\"}]},{\"info\":{\"isRootEntity\":true,\"subClasses\":[]},\"path\":\"model::Trader\",\"properties\":[{\"_type\":\"MappedProperty\",\"mappedPropertyInfo\":{\"multiplicity\":{\"lowerBound\":1,\"upperBound\":1},\"propertyType\":\"String\"},\"name\":\"fullName\"}]}]}",
+                api.analyzeModelCoverage(new MappingModelCoverageAnalysisInput("vX_X_X", "model::mapping", modelContextData), true, true, null).getEntity().toString());
+    }
 
     @Test
     public void testMappingRuntimeCompatibilityAnalysis() throws IOException
